Make the unrolling logic independent of the merge event
